test(cli): move parser tests into a separate module
